backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
duckdb==1.1.3
plotly==5.24.1
google-generativeai==0.8.3
tavily-python==0.5.0
pyyaml==6.0.2
python-dotenv==1.0.1
python-multipart==0.0.18
httpx==0.28.1
